check user inputs

+++ FinalProject/WEB-INF/classes/cast_del.java
@@ -12,10 +12,14 @@ public class cast_del extends HttpServlet
 			Statement state4 = null;
 			ResultSet result = null;
 			ResultSet alertResult = null;      
+			ResultSet alertResult2 = null;      
+			ResultSet alertResult3 = null;      
 			String query1="";        
 			String query2="";        
 			Connection con=null; 
 			String alertQuery="";  
+			String alertQuery2="";  
+			String alertQuery3="";  
           
 			// get parameters
             String c_id = (request.getParameter("CastID")).toString();
@@ -59,18 +63,63 @@ public class cast_del extends HttpServlet
 		// alert query to notify that person has been deleted
 		try 
 		{ 
+			//Queries to check if user entries are valid
 			alertQuery="SELECT FirstName, LastName FROM person where PersonID = '" + p_id + "'";
+			alertQuery2="SELECT CharName FROM cast where CastID = '" + c_id + "'";
+			alertQuery3="SELECT CharName from cast where CastID = '" + c_id + "' and PersonID = '" + p_id + "'";
 			PreparedStatement pstmt1 = con.prepareStatement(alertQuery);
 			alertResult = pstmt1.executeQuery();
+			PreparedStatement pstmt2 = con.prepareStatement(alertQuery2);
+			alertResult2 = pstmt2.executeQuery();
+			PreparedStatement pstmt3 = con.prepareStatement(alertQuery3);
+			alertResult3 = pstmt3.executeQuery();
 
+			// Checking if person_id is valid
+			if(!alertResult.next()){
+				out.println("<script>function showAlertOnLoad() {alert(\"Error: The person ID you entered is not valid.\");}</script>");
+			}
+			// Checking if cast_id is valid
+			else if(!alertResult2.next()){
+				out.println("<script>function showAlertOnLoad() {alert(\"Error: The cast ID you entered is not valid.\");}</script>");
+			}
+			// Checking if person_id and cast_ID are a valid combination
+			else if(!alertResult3.next()){
+				out.println("<script>function showAlertOnLoad() {alert(\"Error: The cast ID and person ID you entered are not a valid combination. Please reference the Cast and Person tables.\");}</script>");
+			}
+			else
+			{
+				//exec query to delete
+				try{
+					// creating delete alert
+					String fname = alertResult.getString("FirstName");
+					String lname = alertResult.getString("LastName");
+					out.println("<script>function showAlertOnLoad() {alert(\"You have deleted one record for the cast member " + fname + " " + lname + " \");}</script>");
+
+					try 
+					{ 
+						// build query
+						query1 = "delete from cast where CastID = '" + c_id + "' and PersonID = '" + p_id + "'";
+						result=state4.executeQuery(query1);
+							
+					}
+					catch (SQLException e) 
+					{
+						System.err.println("SQLException while executing SQL Statement."); 
+					}
+				}
+				catch (SQLException e) 
+				{
+					System.err.println("SQLException while executing SQL Statement."); 
+				}
+			}
+		
 	  	}
 		catch (SQLException e) 
 		{
 			System.err.println("SQLException while executing SQL Statement."); 
 		}
 		
-		// build query
-		query1 = "delete from cast where CastID = '" + c_id + "' and PersonID = '" + p_id + "'";
+		// build query to show table 
 		query2 = "select c.CastID, m.MovieID, m.title, p.PersonID, p.FirstName, p.LastName,  c.CharName, p.PayK, p.PersonType FROM MOVIE m, PERSON p, CAST c WHERE m.MovieID = c.MovieID AND p.personID = c.personID order by c.CastID";
 		
 		//write to html file
@@ -79,20 +128,7 @@ public class cast_del extends HttpServlet
 		out.println("<link rel=\"stylesheet\" href=\"https://fonts.googleapis.com/css?family=Roboto:regular,bold,italic,thin,light,bolditalic,black,medium&amp;lang=en\"> ");
 		out.println("<link rel=\"stylesheet\" type=\"text/css\" href=\"\\FinalProject\\html\\CSS\\base.css\">");
 		
-		//exec query
-		try{
-			while (alertResult.next()) {
-
-				String fname = alertResult.getString("FirstName");
-				String lname = alertResult.getString("LastName");
-				out.println("<script>function showAlertOnLoad() {alert(\"You have deleted one record for the cast member " + fname + " " + lname + " \");}</script>");
-
-			}
-		}
-		catch (SQLException e) 
-		{
-			System.err.println("SQLException while executing SQL Statement."); 
-		}
+		
 
 		//write to html
 		out.println("</head><body onload=\"showAlertOnLoad()\"><br/><br/><br/><br/><br/><br/><br/><section id=\"javaSection\">");
@@ -113,10 +149,9 @@ public class cast_del extends HttpServlet
 		out.println("<th>Person Type</th>");
         out.println("</tr>");
 		
-		//exec query
+		//exec table query
        	try 
 		{ 
-			result=state4.executeQuery(query1);
 			result=state4.executeQuery(query2);
 				
 	  	}
